+++ facade/src/main/java/com/timevale/forward/facade/api/request/TroubleTicketAddReq.java
@@ -135,7 +135,7 @@ public class TroubleTicketAddReq extends BaseReq {
     private Integer cause;
 
     @ApiModelProperty("故障扣分")
-    private Integer deductPoints;
+    private Long deductPoints;
 
     @ApiModelProperty("数据统计说明")
     private String dataStatistics;
