+++ facade/src/main/java/com/timevale/forward/facade/api/result/TroubleTicketDetailVO.java
@@ -168,7 +168,7 @@ public class TroubleTicketDetailVO extends ToString {
     private String causeName;
 
     @ApiModelProperty("故障扣分")
-    private Integer deductPoints;
+    private Long deductPoints;
 
     @ApiModelProperty("数据统计说明")
     private String dataStatistics;
